[SUPC] Regenerated demo for supc_supc_wakeup_wkup_sam_e70_xult MH3-27665

+++ apps/supc/supc_wakeup_wkup/firmware/src/config/sam_e70_xult/peripheral/supc/plib_supc.c
@@ -92,7 +92,7 @@ void SUPC_WaitModeEnter (WAITMODE_FLASH_STATE flash_lpm, WAITMODE_WKUP_SOURCE so
     __enable_irq();
     
     /* FLASH  Low power mode and Wakeup source	*/
-    PMC_REGS->PMC_FSMR = (flash_lpm | source);
+    PMC_REGS->PMC_FSMR = ((uint32_t) flash_lpm | (uint32_t) source);
     
     /* Set Flash Wait State at 0	*/
     EFC_REGS->EEFC_FMR = EEFC_FMR_FWS(0) | EEFC_FMR_CLOE_Msk;
